MVPリリース時の機能について相談
