+++ app/controllers/posts_controller.rb
@@ -19,16 +19,7 @@ class PostsController < ApplicationController
     @current_user = current_user
     @analyse_face_detail = AnalyseFaceDetail.find(@post.analyse_face_detail_id)
     @analyse_face_emotion = AnalyseFaceEmotion.find(@post.analyse_face_emotion_id)
-    @data = {
-      "喜び" => @analyse_face_emotion.happy,
-      "悲しむ" => @analyse_face_emotion.sad,
-      "怒り" => @analyse_face_emotion.angry,
-      "驚き" => @analyse_face_emotion.surprised,
-      "落ち着き" => @analyse_face_emotion.calm,
-      "恐れ" => @analyse_face_emotion.fear,
-      "混乱" => @analyse_face_emotion.confused,
-      "嫌悪" => @analyse_face_emotion.disgusted
-    }
+    @data = @analyse_face_emotion.emotion_data
     @comments = @post.comments.includes(:user).order(created_at: :desc)
     @comment = current_user.comments.new
     @favorite_counts = @post.favorites.count
