+++ app/models/photo.rb
@@ -6,4 +6,23 @@ class Photo < ApplicationRecord
 
   validates :user_id, presence: true
 
+  def s3_image
+    client = Aws::Rekognition::Client.new(region: 'ap-northeast-1')
+    response = client.detect_faces({
+      image: {
+        s3_object: {
+          bucket: "smile-images",
+          name: self.image.key
+        }
+      },
+      attributes: ['ALL']
+    })
+
+    face_detail = response.face_details.first
+  end
+
+  def analysed?
+    self.analyse_face_detail.present? && self.analyse_face_emotion.present?
+  end
+
 end
