+++ app/views/photos/detect_faces.html.erb
@@ -59,7 +59,7 @@
     <% else %>
       <div class="fixed bottom-0  w-full z-50 flex justify-center items-center bg-gray-200 py-2 px-4 shadow-md">
         <%= form.submit '投 稿', class: "bg-white border-2 border-black text-xl px-16 py-2 ml-5 rounded shadow-lg transform transition-transform hover:scale-105" %>
-        <%= link_to "シェア", "https://twitter.com/share?url=#{ request.url }&text=【君の笑顔は...】%0a%0a#{@score}点！", target: '_blank', class: "bg-white border-2 border-black text-xl px-16 py-2 ml-5 rounded shadow-lg transform transition-transform hover:scale-105"  %>
+        <%= link_to "シェア", "https://twitter.com/share?url=#{ request.url }&text=【君の笑顔は...】%0a%0a#{@analyse_face_emotion.score }点！", target: '_blank', class: "bg-white border-2 border-black text-xl px-16 py-2 ml-5 rounded shadow-lg transform transition-transform hover:scale-105"  %>
         <dic class="field flex justify-center item-center ml-10">
           <%= form.check_box :blur, :as => :boolean, class: " mt-1" %>
           <%= form.label :投稿時に顔写真をぼかします, class: "text-xl ml-2" %>
